services/upload_service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic-settings==2.6.1
redis==5.2.1
pydicom==3.0.1
Pillow==11.0.0
numpy==2.1.3
aiofiles==24.1.0
structlog==24.4.0
prometheus-fastapi-instrumentator==7.0.0
python-multipart==0.0.17
